requirements.txt
pytest==6.1.1
Appium-Python-Client==1.1.0
selenium~=3.141.0
python-dotenv~=1.0.0